Give the C boundary a module of its own
`CSignature` is the vocabulary of both FFI directions and of the entry
point the compiler writes, and `CTypeShape` says what a C declaration
carries. Both lived where they were first needed — the export statement
and the Fix type — so a reader looking for "the type of a C function"
had no place to look.

`src/ffi.rs` now holds the C side of the boundary: what a C declaration
says about a value, how a Fix type constructor reaches C, the signature
the two directions have to agree on, and the declaration they meet at.
Moving `TyCon`'s C methods along with the shapes leaves the dependency
one-way, and `ast/types.rs` drops the three inkwell type imports it kept
only for them. Every call site is unchanged: these are inherent methods,
which resolve wherever their impl is written.

Refs #319, #320
