+++ src/ast/export_statement.rs
@@ -1,13 +1,13 @@
 // Export syntax: `FFI_EXPORT[fix_value_name, c_function_name];`
 
 use crate::ast::expr::ExprNode;
-use crate::ast::name::{FullName, Name};
+use crate::ast::name::FullName;
 use crate::ast::program::TypeEnv;
 use crate::ast::types::Scheme;
-use crate::ast::types::{tycon, TyCon, Type, TypeNode};
+use crate::ast::types::{Type, TypeNode};
 use crate::configuration::OutputFileType;
-use crate::constants::{I32_NAME, PTR_NAME, STD_NAME};
 use crate::error::Errors;
+use crate::ffi::{assert_crosses_as_c_type, c_boundary_tycon, CSignature};
 use crate::fixstd::builtin::{make_iostate_ty, run_io};
 use crate::fixstd::runtime::compiler_defined_c_function_reason;
 use crate::generator::Generator;
@@ -16,9 +16,6 @@ use crate::object::create_obj;
 use crate::object::ObjectFieldType;
 use crate::parse::sourcefile::Span;
 use crate::rc_ir::ast::RcState;
-use inkwell::attributes::AttributeLoc;
-use inkwell::types::{BasicMetadataTypeEnum, BasicType};
-use inkwell::values::FunctionValue;
 use std::sync::Arc;
 
 // The export statement.
@@ -191,54 +188,6 @@ impl ExportStatement {
     }
 }
 
-/// Assert that a value of Fix type `ty` travels as the one scalar the C type `c_ty` names, which is
-/// what lets the generated function hand it to C as it stands.
-///
-/// Counting the parts alone would let an aggregate through, since a value too wide to split is
-/// carried as one part holding the whole of it, and C would then be handed a structure whose layout
-/// it classifies by its own rules. `c_boundary_tycon` admits nothing with either shape.
-fn assert_crosses_as_c_type<'c, 'm>(
-    c_ty: &Arc<TyCon>,
-    ty: &Arc<TypeNode>,
-    gc: &mut Generator<'c, 'm>,
-) {
-    let embedded_ty = ty.get_embedded_type(gc);
-    let parts = gc.type_parts(embedded_ty);
-    assert_eq!(
-        parts.len(),
-        1,
-        "`{}` reached an exported signature, where a value has to be one scalar",
-        ty.to_string()
-    );
-    assert_eq!(
-        parts[0],
-        c_ty.get_c_type(gc.context).unwrap(),
-        "`{}` crosses as the C type `{}`, so the two lay it down alike",
-        ty.to_string(),
-        c_ty.to_string()
-    );
-}
-
-/// The C type a value of `ty` crosses the FFI boundary as, and `None` for a value the C ABI cannot
-/// carry the way Fix lays it down.
-///
-/// A value with one scalar — an integer, a floating point number, or a pointer — is laid down
-/// identically by Fix and by C, and a boxed value is a pointer. An aggregate is laid down
-/// differently: the C ABI classifies a structure by its size and by the class of each of its
-/// eightbytes (System V AMD64), or by whether it is a homogeneous floating-point aggregate
-/// (AAPCS64), and the shapes on which that agrees with Fix's element-wise layout differ from target
-/// to target.
-fn c_boundary_tycon(ty: &Arc<TypeNode>, type_env: &TypeEnv) -> Option<Arc<TyCon>> {
-    let head = ty.toplevel_tycon()?;
-    if ty.is_box(type_env) {
-        return Some(tycon(FullName::from_strs(&[STD_NAME], PTR_NAME)));
-    }
-    if !head.is_c_scalar() {
-        return None;
-    }
-    Some(head)
-}
-
 /// The error message for a type the C ABI cannot carry appearing in an exported function's
 /// signature.
 ///
@@ -255,146 +204,6 @@ fn unexportable_type_msg(ty: &Arc<TypeNode>, position: &str) -> String {
         return head + ". Use `U8` or `CInt`, and convert it on the Fix side.";
     }
     head + ". An exported function can exchange scalar values: integers (`I8` to `I64`, `U8` to `U64`), floating point numbers (`F32`, `F64`), and pointers (`Ptr`, and boxed values, which cross as an opaque pointer). The C types in `Std::FFI` such as `CInt` are aliases of these. To exchange a struct, take a `Ptr` to memory the foreign side owns and copy through it with `memcpy`; `Std::FFI::borrow_boxed` and `mutate_boxed` give a pointer to the payload of a boxed value, and `Std::Array::borrow_elements` and `mutate_elements` a pointer to an array's elements."
-}
-
-/// The type of a C function, as the Fix type constructors standing for the C types it exchanges.
-///
-/// A program describes a C function in two ways: an `FFI_CALL` declares one it calls, and an
-/// `FFI_EXPORT` defines one it offers. Both descriptions of one name reach the module as a single
-/// function, through which every call in the program goes, so they have to give one signature.
-pub struct CSignature {
-    /// The type of each parameter, in the order the C function takes them.
-    pub param_tys: Vec<Arc<TyCon>>,
-    /// The type of the result, the unit type where the C function returns nothing.
-    pub ret_tycon: Arc<TyCon>,
-    /// Whether the signature ends in `...`.
-    pub is_var_args: bool,
-}
-
-/// The signature of the entry point the compiler generates: `int32_t (int32_t, void *)`, taking
-/// `argc` and `argv` as the C runtime passes them.
-///
-/// The compiler writes this function's body, so it is the one C function a program describes without
-/// naming it. A program that does name it — an `FFI_CALL` of `main`, which re-runs the program —
-/// gives this signature, and `Program::validate_c_function_calls` reports one that gives another.
-pub fn c_entry_point_signature() -> CSignature {
-    let std_tycon = |name: &str| tycon(FullName::from_strs(&[STD_NAME], name));
-    CSignature {
-        param_tys: vec![std_tycon(I32_NAME), std_tycon(PTR_NAME)],
-        ret_tycon: std_tycon(I32_NAME),
-        is_var_args: false,
-    }
-}
-
-impl CSignature {
-    /// The signature an `FFI_CALL` writes for the function it calls.
-    pub fn of_ffi_call(
-        ret_tycon: &Arc<TyCon>,
-        param_tys: &Vec<Arc<TyCon>>,
-        is_var_args: bool,
-    ) -> CSignature {
-        CSignature {
-            param_tys: param_tys.clone(),
-            ret_tycon: ret_tycon.clone(),
-            is_var_args,
-        }
-    }
-
-    /// The signature of the C function generated for a value exported at `exported_ty`, whose every
-    /// position `ExportedFunctionType::validate` has admitted as one the C ABI can carry.
-    pub fn of_ffi_export(exported_ty: &ExportedFunctionType, type_env: &TypeEnv) -> CSignature {
-        let boundary_tycon = |ty: &Arc<TypeNode>| {
-            c_boundary_tycon(ty, type_env)
-                .unwrap_or_else(|| panic!("`{}` reached an exported signature", ty.to_string()))
-        };
-        CSignature {
-            param_tys: exported_ty.doms.iter().map(boundary_tycon).collect(),
-            ret_tycon: if exported_ty.codom.is_unit() {
-                exported_ty.codom.toplevel_tycon().unwrap()
-            } else {
-                boundary_tycon(&exported_ty.codom)
-            },
-            is_var_args: false,
-        }
-    }
-
-    /// Whether this signature and `other` declare the same C function: the two describe every
-    /// position the same way, down to what a declaration of it carries — `CTypeShape` holds which
-    /// differences between two Fix types that is, and which it is not.
-    pub fn agrees_with(&self, other: &CSignature) -> bool {
-        self.is_var_args == other.is_var_args
-            && self.param_tys.len() == other.param_tys.len()
-            && self.ret_tycon.c_type_shape() == other.ret_tycon.c_type_shape()
-            && (self.param_tys.iter())
-                .zip(other.param_tys.iter())
-                .all(|(a, b)| a.c_type_shape() == b.c_type_shape())
-    }
-
-    /// The function `name` of this signature in the module, declaring it where nothing declares it
-    /// yet. Every description of one C name goes through here, which is what puts the calls a
-    /// program makes and the definition it exports on one function.
-    pub fn get_or_declare_in_module<'c, 'm>(
-        &self,
-        name: &Name,
-        gc: &Generator<'c, 'm>,
-    ) -> FunctionValue<'c> {
-        if let Some(declared) = gc.module.get_function(name) {
-            return declared;
-        }
-        let param_c_tys = self
-            .param_tys
-            .iter()
-            .map(|param_ty| {
-                // A parameter of type `()` is rejected where the signature is written: `void` is a
-                // result alone.
-                param_ty.get_c_type(gc.context).unwrap().into()
-            })
-            .collect::<Vec<BasicMetadataTypeEnum>>();
-        let fn_ty = match self.ret_tycon.get_c_type(gc.context) {
-            None => gc
-                .context
-                .void_type()
-                .fn_type(&param_c_tys, self.is_var_args),
-            Some(ret_c_ty) => ret_c_ty.fn_type(&param_c_tys, self.is_var_args),
-        };
-        let func = gc.module.add_function(name, fn_ty, None);
-        assert_eq!(
-            func.get_name().to_str().unwrap(),
-            name,
-            "the C function enters the module under the name it was given"
-        );
-        gc.add_c_integer_extension_attribute(func, AttributeLoc::Return, &self.ret_tycon);
-        for (i, param_ty) in self.param_tys.iter().enumerate() {
-            gc.add_c_integer_extension_attribute(func, AttributeLoc::Param(i as u32), param_ty);
-        }
-        func
-    }
-
-    /// The signature as a C declaration of the function `name` reads.
-    ///
-    /// # Examples
-    /// A function of two arguments reads `int32_t c_pick(int32_t, int32_t)`, one of none
-    /// `void c_now(void)`, and a variadic one `int32_t c_report(void *, ...)`.
-    pub fn declaration_of(&self, name: &Name) -> String {
-        let mut params = self
-            .param_tys
-            .iter()
-            .map(|param_ty| param_ty.c_type_name())
-            .collect::<Vec<_>>();
-        if self.is_var_args {
-            params.push("...".to_string());
-        }
-        // C writes the empty parameter list of a declaration as `(void)`.
-        if params.is_empty() {
-            params.push("void".to_string());
-        }
-        format!(
-            "{} {}({})",
-            self.ret_tycon.c_type_name(),
-            name,
-            params.join(", ")
-        )
-    }
 }
 
 // The type of an exported Fix value, split into the parts the generated C function is built from.
